backend/requirements.txt
google-adk
Flask
Flask-Cors
uvicorn
pydantic
python-dotenv
google-generativeai